V 1.0.6

+++ src/main/java/org/example/NumberLenghtMod.java
@@ -2,9 +2,9 @@ package org.example;
 
 public class NumberLenghtMod {
 
-    public int sumDigits(int number) {
+    public double sumDigits(int number) {
         int cont = 0;
-        int sum = 0;
+        double sum = 0;
         while (number > 0) {
             sum += number % 10;
             number /= 10;
